Tests structure improvements.
