+++ src/test/scala/org/ergoplatform/nodeView/mempool/MempoolPerformanceBench.scala
@@ -1,7 +1,6 @@
-package org.ergoplatform
+package org.ergoplatform.nodeView.mempool
 
 import org.ergoplatform.modifiers.mempool.ErgoTransaction
-import org.ergoplatform.nodeView.mempool.ErgoMemPool
 import org.ergoplatform.utils.generators.{ErgoGenerators, ErgoTransactionGenerators}
 import org.scalacheck.Gen
 import org.scalatest.PropSpec
